get simple renegotiation working biggest part of the change was duplicating state in conn params for read and write states, and tracking two connparams objects simultaneously--the current parameters used to decrypt incoming messages and the "next" parameters, storing state to compute the secrets for the next negotiation
symmetric cipherer needs to use a shared ptr to avoid clearing state in keyandprov

initializing cipher parameters and cipherer objects is done in a callback now, which is called on every new clienthello, basically.

there are still a lot of places where the code is reading the wrong endpoint's state, such as the "read version" when it should be using the "write version", which is incorrect between reading the client's changecipherspec and the server sending a changecipherspec. consequently, the renegotiation only currently supports renegotiating exactly the same parameters.

remove the stupid block of code that re-parses all the messages looking for handshakes and saves them. instead, whenever we create a handshake message we'll manually add it to the list of handshake messages

move master secret computation into connection parameters

oops, forgot to error check after encryption and decryption. oops.

in the exe, get to move a bunch of crypto variables off of the server object itself and into locals.